Accentuation

PolytonicGreek.rmaccents — Function

Remove all accent characters from s.

PolytonicGreek.accentword — Function

Accent word according to specified system of accent placement. wrd is a string value representing a single lexical token. placement is one of :RECESSIVE for recessive accent or :PENULT for persistent accent on the penultimate syllable. Note that it is not possible to accent the ultima correctly without additional morphological information beyond the string value of the token.

Examples:

PolytonicGreek.allaccents — Function

List all code points including an accent.

PolytonicGreek.accentstripdict — Function

Dictionary mapping Unicode codepoints with accents to codepoint with accent stripped.

We can't just use Unicode.normalize with stripmark flag because that strips breathings, iota subscripts, and diaereses as well.

PolytonicGreek.flipdict — Function

Flip dictionary

PolytonicGreek.acutedict — Function

Dictionary of acutes.

PolytonicGreek.circumflexdict — Function

Dictionary of circumflexes

PolytonicGreek.addacute — Function

Add an acute accent to a single vowel or diphthong.

julia> PolytonicGreek.addacute("ᾀ")
"ᾄ"
julia> PolytonicGreek.addacute("τα")
┌ Warning: addacute: can't add acute accent to vowel τα

PolytonicGreek.addcircumflex — Function

Add a circumflex accent to a single vowel or diphthong

Examples:

julia> PolytonicGreek.addcircumflex("ᾀ")
"ᾆ"
julia> PolytonicGreek.addcircumflex("τα")
┌ Warning: addcircumflex: can't add circumflex accent to vowel τα

PolytonicGreek.accentsyllable — Function

Add specified accent to a single syllable. syll is a string value for a single syllable. accent is either :ACUTE or :CIRCUMFLEX. The function returns nothing for any other symble for accent.

julia> PolytonicGreek.accentsyllable("των", :CIRCUMFLEX)
"τῶν"
julia> PolytonicGreek.accentsyllable("τα", :ACUTE)
"τά"
julia> PolytonicGreek.accentsyllable("ᾀ", :ACUTE)
"ᾄ"

PolytonicGreek.accentultima — Function

Place accent on ultima

PolytonicGreek.accentpenult — Function

Place accent on penult

PolytonicGreek.accentantepenult — Function

Place accent on antepenult

PolytonicGreek.ultima — Function

Return ultima

PolytonicGreek.penult — Function

Return penult

PolytonicGreek.longsyllable — Function

True if syll is metrically long by nature.

Examples:

julia> PolytonicGreek.longsyllable("τει")
true
julia> PolytonicGreek.longsyllable("τε")
false

PolytonicGreek.shortsyllable — Function

True if syll is not long by nature. Examples:

julia> PolytonicGreek.shortsyllable("ε")
true
julia> PolytonicGreek.shortsyllable("ει")
false

PolytonicGreek.includesdiphthong — Function

True if s contains a diphthong.

Examples:

julia> PolytonicGreek.includesdiphthong("εὐθύς")
true
julia> PolytonicGreek.includesdiphthong("ᾔα")
false

PolytonicGreek.flipaccent — Function

Convert grave accent to acute.

PolytonicGreek.vowelsonly — Function

Remove all consonants from s.

Example:

julia> PolytonicGreek.vowelsonly("τῶν")
"ῶ"

PolytonicGreek.tokenform — Function

tokenaccent(s::AbstractString)

PolytonicGreek.countaccents — Function

PolytonicGreek.stripenclitic — Function

Remove any second enclitic accent from s.

PolytonicGreek.normalizediaresis — Function

Create NFKC forms of code points with diaeresis.

PolytonicGreek.normalizeaccentcombos — Function

Collect a normalized set of all codepoints with accent.