Orthography

PolytonicGreek.LiteraryGreekOrthography — Type

An orthographic system for standard orthography of printed literary Greek.

PolytonicGreek.literaryGreek — Function

Instantiate a LiteraryGreekOrthography with correct code points and token types.

literaryGreek()

PolytonicGreek.tokenizeLiteraryGreek — Function

Tokenize a string in standard literary Greek orthography.

tokenizeLiteraryGreek(s)

Internals: orthography

PolytonicGreek.alphabetic — Function

Compose a string with all alphabetic characters.

PolytonicGreek.punctuation — Function

Compose a string with all alphabetic characters.

PolytonicGreek.isAlphabetic — Function

True if all characters in s are alphabetic.

isAlphabetic(s)

PolytonicGreek.isPunctuation — Function

True if all characters in s are puncutation.

isPunctuation(s)

PolytonicGreek.sortWords — Function

Alphabetically sort a list of words in Unicode Greek.

sortWords(words)

PolytonicGreek.tokenforstring — Function

Create correct OrthographicToken for a given string.

tokenforstring(s)

PolytonicGreek.splitPunctuation — Function

Split off any trailing punctuation and return an Array of leading string + trailing punctuation.

splitPunctuation(s)