API documentation

String composition

PolytonicGreek.strcat — Function

Concatenate s1 and a list of strings slist, taking account of any phonological modifications required by orthography ortho.

strcat(ortho, s1, slist)

Concatenate the pair of strings s1 and s2, taking account of any phonological modifications required by orthography ortho.

strcat(s1, s2, ortho)

Concatenate s1 and s2 d, taking account of necessary phonological modifications.

strcat(s1, s2, ortho)

PolytonicGreek.sortWords — Function

Alphabetically sort a list of words in Unicode Greek.

sortWords(words, ortho)

Phonology

PolytonicGreek.vocalic — Function

True if s is a vowel or diphthong.

vocalic(s)

True if c is a vowel or diphthong.

vocalic(c)

PolytonicGreek.consonantal — Function

True if s is a consonant.

consonantal(s)

PolytonicGreek.liquids — Function

Identify code points representing liquid consonants.

List code points for liquid consonants in literary Greek.

liquids(ortho)

PolytonicGreek.labials — Function

Identify code points representing labial consonants.

List code points for labial consonants in literary Greek.

labials(ortho)

PolytonicGreek.dentals — Function

Identify code points representing dental consonants.

List code points for dental consonants in literary Greek.

dentals(ortho)

PolytonicGreek.palatals — Function

Identify code points representing palatal consonants.

List code points for palatal consonants in literary Greek.

palatals(ortho)

PolytonicGreek.stops — Function

Identify code points representing stops.

Reduplication and augment

PolytonicGreek.augment — Function

Generic method of function to augment a string.

Parameters

ortho An instance of a GreekOrthography

Implementatiοn of GreekOrthography's augment function for literary Greek.

augment(s, ortho)

NB: augment removes all accents from the resulting string.

Parameters

ortho An instance of a LiteraryGreekOrthography

s An optional string to augment. If is not included, the function returns

a default augment string that can be applied to verb forms starting with a consonant (except note that ρ doubles after augment in standard literary Greek orthography).

PolytonicGreek.augment_initial — Function

Generic method of function to identify string for syllabic augment in initial position.

Parameters

ortho An instance of a GreekOrthography

Identify string to use for syllabic augment in word-initial position.

augment_initial(ortho)

PolytonicGreek.augment_medial — Function

Generic method of function to identify string for syllabic augment when preceded by a prefix.

Parameters

ortho An instance of a GreekOrthography

Identify string to use for syllabic augment in compound verb.

augment_medial(ortho)

PolytonicGreek.reduplicate — Function

Add reduplications to s, taking account of any phonological modifications required by orthography ortho.

reduplicate(s, ortho)

Reduplicate s, taking account of any necessary phonological modifications.

reduplicate(s, ortho)

Accentuation

PolytonicGreek.rmaccents — Function

Remove all accent characters from s.

Parameters

s is a Greek string

ortho is an implementation of GreekOrthography

Default to using literary Greek orthography for rmaccents.

rmaccents(s)

Implement GreekOrthography's rmaccents function for LiteraryGreekOrthography.

rmaccents(s, ortho)

PolytonicGreek.accentword — Function

Accent word s.

accentword(s, placement, ortho)

Parameters

s is a Greek string

placement is :RECESSIVE, or :PENULT

ortho is an implementation of GreekOrthography

Accent word according to specified system of accent placement.

accentword(wrd, placement)
accentword(wrd, placement, ortho)

Parameters

wrd is a string value representing a single lexical token.

placement is one of :RECESSIVE for recessive accent

or :PENULT for persistent accent on the penultimate syllable.

Note that it is not possible to accent the ultima correctly without additional morphological information beyond the string value of the token.

PolytonicGreek.accented — Function

True if s has at least one accent.

accented(s)

PolytonicGreek.proclitics — Function

List proclitics in this orthography.

List proclitics in LiteraryGreekOrthography.

proclitics(ortho)

Based on Smyth ¶ 179.

PolytonicGreek.enclitics — Function

List enclitics in this orthography.

List enclitics in LiteraryGreekOrthography.

enclitics(ortho)

Based on Smyth ¶ 181.

PolytonicGreek.rmbreathing — Function

Remove any breathing from s.

rmbreathing(s, ortho)

PolytonicGreek.addbreathing — Function

Add the specified breathing to s.

addbreathing(s, breath, ortho)

breath should be one of :SMOOTH or :ROUGH.

PolytonicGreek.countaccents — Function

Count accents in s.

Parameters

s is a Greek string

ortho is an implementation of GreekOrthography

Default to using literary Greek orthography for countaccents.

countaccents(s)

PolytonicGreek.accentpenult — Function

Accent penultimate syllable of wrd.

accentpenult(wrd, accent, ortho)

Parameters

s is a Greek string

accent is :ACUTE, or :CIRCUMFLEX

ortho is an implementation of GreekOrthography

Place accent on penult.

accentpenult(wrd, accent)
accentpenult(wrd, accent, ortho)

PolytonicGreek.accentultima — Function

Accent final syllable of wrd.

Parameters

s is a Greek string

accent is :ACUTE, or :CIRCUMFLEX

ortho is an implementation of GreekOrthography

Place accent on ultima.

accentultima(wrd, accent)
accentultima(wrd, accent, ortho)

Syllabification

PolytonicGreek.syllabify — Function

Generic method of GreekOrthography's syllabify function.

syllabify(s, ortho)

Allow syllabification defaulting to LiteraryGreekOrthography.

syllabify(s)

Split string s into an Array of strings representing syllables.

syllabify(s, ortho)

Example

syllables = PolytonicGreek.syllabify("κελεύει")
join(syllables, "-")
"κε-λευ-ει"

Orthographic system

PolytonicGreek.LiteraryGreekOrthography — Type

An orthographic system for standard orthography of printed literary Greek.

PolytonicGreek.literaryGreek — Function

Instantiate a LiteraryGreekOrthography with correct code points and token types.

literaryGreek()

Internals

Syllabification is implemented with a chain of functions applying regular expressions to an input string. These internal functions would not normally be of any use or interest to anyone who simply wants to break a string in polytonic Greek into syllables.

PolytonicGreek.splitmorphemeboundary — Function

Respect explicitly indicated morpheme boundaries.

splitmorphemeboundary(s)

Example

PolytonicGreek.splitmorphemeboundary("εἰσ#ῄα")
"εἰσ ῄα"

PolytonicGreek.splitmunu — Function

Mu+nu stay together.

splitmunu(s)

ἀναμιμνησκόμενος breaks as "ἀναμι μνησκομενος"

PolytonicGreek.splitliqcons — Function

Split between a liquid and non-liquid consonant.

PolytonicGreek.splitdiphthongvowel — Function

splitdiphthongvowel(s)

Split between a diphthong and following vowel.

Example

PolytonicGreek.splitdiphthongvowel("κελεύει")
"κελευ ει"

PolytonicGreek.splitvoweldiphthong — Function

Split between a vowel and a following diphthong.

"οἰκι αις"

Internals

PolytonicGreek.alphabetic — Function

Compose a string with all alphabetic characters.

alphabetic()

PolytonicGreek.punctuation — Function

Compose a string with all punctuation characters.

punctuation()

PolytonicGreek.isAlphabetic — Function

True if all characters in s are alphabetic.

isAlphabetic(s)

PolytonicGreek.isPunctuation — Function

True if all characters in s are puncutation.

isPunctuation(s)

Missing docstring.

Missing docstring for PolytonicGreek.sortWords. Check Documenter's build log for details.

PolytonicGreek.tokenforstring — Function

Create correct OrthographicToken for a given string.

tokenforstring(s)

PolytonicGreek.splitPunctuation — Function

Split off any trailing punctuation and return an Array of leading string + trailing punctuation.

splitPunctuation(s)

Internals

PolytonicGreek.allaccents — Function

List all code points including an accent.

allaccents()

Default to using LiteraryGreekOrthography.

List all code points including an accent in LiteraryGreekOrthography.

allaccents(ortho)

PolytonicGreek.accentstripdict — Function

Dictionary mapping Unicode codepoints with accents to codepoint with accent stripped.

accentstripdict()

Default to using LiteraryGreekOrthography.

Dictionary mapping Unicode codepoints with accents to codepoint with accent stripped for LiteraryGreekOrthography.

accentstripdict(ortho)

We can't just use Unicode.normalize with stripmark flag because that strips breathings, iota subscripts, and diaereses as well.

PolytonicGreek.flipdict — Function

Dictionary of barytone to oxytone conversions.

flipdict()

Default to using LiteraryGreekOrthography.

Dictionary of barytone to oxytone conversions for LiteraryGreekOrthography.

flipdict(ortho)

PolytonicGreek.acutedict — Function

Dictionary of acutes.

acutedict()

Default to using LiteraryGreekOrthography.

Dictionary of acutes for LiteraryGreekOrthography.

acutedict(ortho)

PolytonicGreek.circumflexdict — Function

Dictionary of circumflexes.

circumflexdict()

Default to using LiteraryGreekOrthography.

Dictionary of circumflexes for LiteraryGreekOrthography.

circumflexdict(ortho)

PolytonicGreek.addacute — Function

Add an acute accent to a single vowel or diphthong.

addacute(vowel)
addacute(vowel, ortho)

Examples

julia> PolytonicGreek.addacute("ᾀ")
"ᾄ"
julia> PolytonicGreek.addacute("τα")
┌ Warning: addacute: can't add acute accent to vowel τα

PolytonicGreek.addcircumflex — Function

Add a circumflex accent to a single vowel or diphthong

addcircumflex(vowel)
addcircumflex(vowel, ortho)

Examples

julia> PolytonicGreek.addcircumflex("ᾀ")
"ᾆ"
julia> PolytonicGreek.addcircumflex("τα")
┌ Warning: addcircumflex: can't add circumflex accent to vowel τα

PolytonicGreek.accentsyllable — Function

Add specified accent to a single syllable.

accentsyllable(syll, accent)
accentsyllable(syll, accent, ortho)

Parameters

syll is a string value for a single syllable. accent is either :ACUTE or :CIRCUMFLEX. The function returns nothing for any other symble for accent.

Examples

julia> PolytonicGreek.accentsyllable("των", :CIRCUMFLEX)
"τῶν"
julia> PolytonicGreek.accentsyllable("τα", :ACUTE)
"τά"
julia> PolytonicGreek.accentsyllable("ᾀ", :ACUTE)
"ᾄ"

Missing docstring.

Missing docstring for PolytonicGreek.accentultima. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PolytonicGreek.accentpenult. Check Documenter's build log for details.

PolytonicGreek.accentantepenult — Function

Place default accent(:ACUTE) on antepenult.

accentantepenult(wrd, ortho)

PolytonicGreek.ultima — Function

Return ultima.

ultima(s, ortho)

PolytonicGreek.penult — Function

Return penult.

penult(s, ortho)

PolytonicGreek.longsyllable — Function

True if syll is metrically long by nature.

longsyllable(syll, ortho)

Arguments

syll is text for a single syllable

ortho is a GreekOrthography

Examples

julia> PolytonicGreek.longsyllable("τει")
true
julia> PolytonicGreek.longsyllable("τε")
false

PolytonicGreek.shortsyllable — Function

True if syll is not long by nature.

shortsyllable(syll, ortho)

Examples

julia> PolytonicGreek.shortsyllable("ε")
true
julia> PolytonicGreek.shortsyllable("ει")
false

PolytonicGreek.includesdiphthong — Function

True if s contains a diphthong.

includesdiphthong(s)

Examples

julia> PolytonicGreek.includesdiphthong("εὐθύς")
true
julia> PolytonicGreek.includesdiphthong("ᾔα")
false

PolytonicGreek.flipaccent — Function

Convert grave accent to acute.

flipaccent(s, ortho)

PolytonicGreek.vowelsonly — Function

Remove all consonants from s.

vowelsonly(s, ortho)

Example

julia> PolytonicGreek.vowelsonly("τῶν")
"ῶ"

PolytonicGreek.tokenform — Function

Missing docstring.

Missing docstring for PolytonicGreek.countaccents. Check Documenter's build log for details.

PolytonicGreek.stripenclitic — Function

Remove any second enclitic accent from s.

PolytonicGreek.normalizediaresis — Function

Create NFKC forms of code points with diaeresis.

normalizediaresis()

PolytonicGreek.normalizeaccentcombos — Function

Collect a normalized set of all codepoints with accent.

normalizeaccentcombos()